+++ quicksort_compare/utils.h
@@ -1,7 +1,7 @@
 #ifndef UTILS_H
 #define UTILS_H
 
-int* generate_random_instance(int size, float min_value, float max_value);
+int* generate_random_instance(int size, int min_value, int max_value);
 void print_vector(int* vector, int size);
 int* copy_vector(int* vector, int size);
 int generate_random_number(int min_value, int max_value);
